requirements.txt
# ---- Core ----
streamlit>=1.37.0
pandas>=2.2.2
numpy>=1.26.4

# ---- Excel reading/writing ----
openpyxl>=3.1.2
xlrd>=2.0.1
xlsxwriter>=3.2.0

# ---- Google Sheets integration ----
gspread>=6.1.2
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0

# ---- JSON / config helpers ----
pyyaml>=6.0.2

# ---- Markdown formatting (optional UI niceties) ----
markdown-it-py>=3.0.0

# ---- Date/time helpers ----
python-dateutil>=2.9.0.post0

# ---- (Optional) data cleaning & dedupe helpers ----
fuzzywuzzy[speedup]>=0.18.0

# ---- PDF export (if you use the PDF generator features) ----
reportlab>=3.6

# ---- (Optional) graph/visualization helpers for Visualizer/Validation ----
# Not strictly required for v6.3.0, but handy if you add graph-based views.
networkx>=3.3
pyvis>=0.3.2
